Add files via upload
added color to my "world" screen, then fixed my dog so it could fit in the screen because I found it was too big. Began researching how I could implement a "draw_border" function that will create a border around the image while also utilizing the random module to create random colors for my border. The border will be contain squares as that is what I believe will be easier to code into the function.