Add files via upload
imported random module from Python and created a function that will draw a border around the dog image. I also initialized colors to a list of different colors so that "t" the turtle could create a random color with each border created. 

Part 4 Draw a border (15 pts)
Now lets draw a border of shapes around our picture. I chose to use squares but you can
use any shape you want as long as you use for loops while creating the border. It is much
easier to do one side at a time.
You will need a function call draw_border that takes a turtle and depending on the shape
you choose a size. This function is very specific to your drawing so it will be different for
each person. 

Part 5: Challenge (5pts)
As a challenge see if you can get each shape to draw a random color.
Hint: you will need to import the random function and change the turtle’s color mode